setting view My Team with NO data and fixed Card component on children
